fix: separated hooks to initializing chart
